+++ src/main/java/system/User.java
@@ -8,11 +8,11 @@ public class User {
     private final int id;
     private final String firstName;
     private final String lastName;
-    private int balance;
+    private double balance;
     private final Map<Product, Integer> boughtProducts;
 
 
-    public User(int id, String firstName, String lastName, int balance) {
+    public User(int id, String firstName, String lastName, double balance) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
@@ -21,12 +21,12 @@ public class User {
     }
 
 
-    public void setBalance(int balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
 
 
-    public int getBalance() {
+    public double getBalance() {
         return balance;
     }
